Move GeneratedObject to Common
