+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/FieldSpecSource.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.fieldspecs;
 
 import com.scottlogic.deg.common.profile.constraints.atomic.AtomicConstraint;
-import com.scottlogic.deg.generator.DataBagValueSource;
+import com.scottlogic.deg.common.output.DataBagValueSource;
 
 import java.util.Collection;
 import java.util.Collections;
