+++ generator/src/main/java/com/scottlogic/deg/generator/generation/DataGenerator.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.generation;
 
 import com.scottlogic.deg.common.profile.Profile;
-import com.scottlogic.deg.generator.outputs.GeneratedObject;
+import com.scottlogic.deg.common.output.GeneratedObject;
 
 import java.util.stream.Stream;
