+++ generator/src/main/java/com/scottlogic/deg/generator/generation/MessagePrintingDataGeneratorMonitor.java
@@ -2,7 +2,6 @@ package com.scottlogic.deg.generator.generation;
 
 import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
-import com.scottlogic.deg.generator.fieldspecs.RowSpec;
 import com.scottlogic.deg.generator.walker.reductive.ReductiveState;
 
 import java.io.PrintWriter;
@@ -28,7 +27,7 @@ public class MessagePrintingDataGeneratorMonitor implements ReductiveDataGenerat
     }
 
     @Override
-    public void rowSpecEmitted(RowSpec rowSpec) {
+    public void rowEmitted() {
         println("RowSpec emitted");
     }
