+++ generator/src/main/java/com/scottlogic/deg/generator/generation/ReductiveDataGeneratorMonitor.java
@@ -8,7 +8,7 @@ import com.scottlogic.deg.generator.walker.reductive.ReductiveState;
 import java.util.Map;
 
 public interface ReductiveDataGeneratorMonitor extends DataGeneratorMonitor {
-    default void rowSpecEmitted(RowSpec rowSpec) {}
+    default void rowEmitted() {}
     default void fieldFixedToValue(Field field, Object current) {}
     default void unableToStepFurther(ReductiveState reductiveState) {}
     default void noValuesForField(ReductiveState reductiveState, Field field) {}
