+++ generator/src/main/java/com/scottlogic/deg/generator/generation/VelocityMonitor.java
@@ -1,7 +1,6 @@
 package com.scottlogic.deg.generator.generation;
 
 import com.google.inject.Inject;
-import com.scottlogic.deg.generator.outputs.GeneratedObject;
 
 import java.io.PrintWriter;
 import java.math.BigDecimal;
@@ -53,7 +52,7 @@ public class VelocityMonitor implements ReductiveDataGeneratorMonitor {
     }
     
     @Override
-    public void rowEmitted(GeneratedObject row) {
+    public void rowEmitted() {
         rowsSinceLastSample++;
         rowsEmitted = rowsEmitted.add(BigInteger.ONE);
     }
