+++ generator/src/main/java/com/scottlogic/deg/generator/utils/DataBagValueIterator.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.utils;
 
-import com.scottlogic.deg.generator.generation.databags.DataBagValue;
-import com.scottlogic.deg.generator.DataBagValueSource;
+import com.scottlogic.deg.common.output.DataBagValue;
+import com.scottlogic.deg.common.output.DataBagValueSource;
 
 import java.util.Iterator;
