+++ generator/src/main/java/com/scottlogic/deg/generator/walker/reductive/ReductiveState.java
@@ -2,9 +2,8 @@ package com.scottlogic.deg.generator.walker.reductive;
 
 import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.common.profile.ProfileFields;
-import com.scottlogic.deg.generator.DataBagValueSource;
 import com.scottlogic.deg.generator.generation.databags.DataBag;
-import com.scottlogic.deg.generator.generation.databags.DataBagValue;
+import com.scottlogic.deg.common.output.DataBagValue;
 import com.scottlogic.deg.generator.walker.reductive.fieldselectionstrategy.FieldValue;
 
 import java.util.*;
